CheckException: Handle CheckRunner Exceptions
These are the parent classes to handle Exception in a RecordCheckRunner
Exceptions can have a Callback Type which makes the overrideable, else
they will just contain an error.
All Callback Exceptions implement the GUIExceptionHandler Interface,
which has the handleGUI method. This method either shows an error to the
User, or it will show the User different options.
One of these can lead to overriding the Exception in the HashMap. The
CheckRunner will skip this Exception in the next run.
Otherwise it will abort and maybe requestFocus on some GUI Fields.