CheckRunner: Add Base Class
This class can be extended for every Record Type that should be checked
before saving.

It has a HashMap where some Exception thrown from the conversion can be
overridden. Each Exception has a static final String to identify it,
this String marks the position in the HashMap where it can be overridden.

Default is to set the Map Field to 1, but other options can be defined.
isOverridden checks if a field/exception has been overridden.

isAllowed checks if the provided AdminRecord is capable of writing to
some parts of the storage.

record() and storage() can be overridden to typecast the DataRecord
and StorageObject objects.