Refactor checks
Main part of this refactor. All checks have been moved from the
EfaBaseFrame to LogbookCheckRunner. The checkRunner gets called as long
as it finishes or a exception's handleGUI method returns false, which
means the error can't be overriden and the user has to change the
input.

This is still a work in progress, some things still behave different
than before and need some adjustments.
But it compiles again! And the basics seem to work...