Experiment with RecordTemplates
This would help to get rid of the special (de)serializer for every
record type. The templates can be annotated for Jackson/JAXB and the
Framework will take care of Serialization.
The respective storageClasses will need a to/fromTemplate method to
interact with the template.

We still need to convert between templates and records, but the
templates can be handled by multiple frameworks and aren't specific to
Jackson, like the Serializers are...